+++ inst/templates/stopwords.R
@@ -1,0 +1,10 @@
+#' Stopwords
+#'
+#' @description
+#' Return a character vector of stopwords.
+#' See \code{stopwords::\link[stopwords:stopwords]{stopwords()}} for details.
+#' @usage stopwords(language = "en", source = "snowball")
+#' @name stopwords
+#' @importFrom stopwords stopwords
+#' @export
+NULL